+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Release History
 ### Pending
 
 #### Update:
+- feat: add [SEP-45](https://github.com/stellar/stellar-protocol/blob/master/ecosystem/sep-0045.md) (Stellar Web Authentication for Contract Accounts) support. Check `stellar_sdk.sep.stellar_soroban_web_authentication` for more details.
 - fix: remove `diagnostic_events_xdr` from `stellar_sdk.soroban_rpc.Events`, use `stellar_sdk.soroban_rpc.GetTransactionResponse.diagnostic_events_xdr` and `stellar_sdk.soroban_rpc.Transaction.diagnostic_events_xdr` instead. ([#1086](https://github.com/StellarCN/py-stellar-base/pull/1086))
 - chore: drop support for Python 3.9.